Backend/functions/requirements.txt
firebase_functions~=0.1.0
google-cloud-aiplatform~=1.48.0
google-generativeai~=0.5.2 